Move webhook specific logic to its own file (discord)
